+++ src/types/ant.ts
@@ -50,6 +50,8 @@ export const IntegerStringSchema = z
     { message: 'Must be a non negative integer string' },
   );
 
+export const AntDescriptionSchema = z.string(); // TODO: add specific limits for description ie max length
+export const AntKeywordsSchema = z.array(z.string()); // TODO: add specific limits for keywords ie max amount and max length
 export const AntRecordSchema = z.object({
   transactionId: ArweaveTxIdSchema.describe('The Target ID of the undername'),
   ttlSeconds: z.number(),
@@ -68,6 +70,8 @@ export const AntBalancesSchema = z.record(
 export const AntStateSchema = z.object({
   Name: z.string().describe('The name of the ANT.'),
   Ticker: z.string().describe('The ticker symbol for the ANT.'),
+  Description: z.string().describe('The description for the ANT.'),
+  Keywords: AntKeywordsSchema.describe('The keywords for the ANT.'),
   Denomination: z
     .number()
     .describe(
@@ -117,12 +121,25 @@ export const AntWriteHandlers = [
   '_eval',
   '_default',
   'transfer',
+  'balance',
+  'balances',
+  'totalSupply',
+  'info',
   'addController',
   'removeController',
+  'controllers',
   'setRecord',
   'removeRecord',
-  'setTicker',
+  'record',
+  'records',
   'setName',
+  'setTicker',
+  'setDescription',
+  'setKeywords',
+  'initializeState',
+  'state',
+  'releaseName',
+  'reassignName',
 ] as const;
 
 export type AoANTWriteHandler = (typeof AntWriteHandlers)[number];
@@ -150,6 +167,9 @@ export const AntInfoSchema = z.object({
   ['Total-Supply']: IntegerStringSchema.describe(
     'Total supply of the ANT in circulation.',
   ),
+  Description: AntDescriptionSchema.describe('The description for the ANT.'),
+  Keywords: AntKeywordsSchema.describe('The keywords for the ANT.'),
+
   Logo: ArweaveTxIdSchema.describe('Transaction ID of the ANT logo.'),
   Denomination: IntegerStringSchema.describe(
     'The number of decimal places to use for the ANT. Defaults to 0 if not set representing whole numbers.',
@@ -234,8 +254,28 @@ export interface AoANTWrite extends AoANTRead {
     { ticker }: { ticker: string },
     options?: WriteOptions,
   ): Promise<AoMessageResult>;
+  setDescription(
+    { description }: { description: string },
+    options?: WriteOptions,
+  ): Promise<AoMessageResult>;
+  setKeywords(
+    { keywords }: { keywords: string[] },
+    options?: WriteOptions,
+  ): Promise<AoMessageResult>;
   setName(
     { name }: { name: string },
+    options?: WriteOptions,
+  ): Promise<AoMessageResult>;
+  releaseName(
+    { name, ioProcessId }: { name: string; ioProcessId: string },
+    options?: WriteOptions,
+  ): Promise<AoMessageResult>;
+  reassignName(
+    {
+      name,
+      ioProcessId,
+      antProcessId,
+    }: { name: string; ioProcessId: string; antProcessId: string },
     options?: WriteOptions,
   ): Promise<AoMessageResult>;
 }